simplifies terraform var passing
rather than building a var list during apply and destroy, we can create
a "terraform.tfvars" that will be used. additionally, this makes running
the terraform steps manually easier when debugging

using hclwrite in case we have more complex variables to write for
aks/eks. we can remove this is the need does not present itself